# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/preview/{shorturl}": {
            "get": {
                "description": "Returns information about a short URL, such as the original URL and metadata.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL Shortener"
                ],
                "summary": "Preview a short URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short URL code",
                        "name": "shorturl",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.URL"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/shorten": {
            "post": {
                "description": "Accepts a long URL, a custom alias, and an optional expiration date, and returns a shortened version",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL Shortener"
                ],
                "summary": "Shortens a URL",
                "parameters": [
                    {
                        "description": "Request body containing URL, custom alias, and expiration date",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.URL"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/{shorturl}": {
            "get": {
                "description": "Finds the original URL from the shortened key and redirects",
                "tags": [
                    "URL Shortener"
                ],
                "summary": "Redirects to the original URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shortened URL key",
                        "name": "shorturl",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.URL": {
            "type": "object",
            "required": [
                "originalURL"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "customURL": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                },
                "expirationDate": {
                    "type": "string"
                },
                "originalURL": {
                    "type": "string"
                },
                "shortURL": {
                    "type": "string"
                }
            }
        }
    }
}